requirements.txt
# requirements.txt
requests
beautifulsoup4
pandas
gspread
gspread-dataframe
google-auth
google-auth-oauthlib